fixes in AquaSimFama
Attributes has been added to configure the data packet size and the destination
of the RTS for a given data packet. See the RTSToNextHop and DataPacketSize
attributes's description for more info.
The initialization of some internal attributes such as m_RTSTxTime and m_maxPropDelay has
been scheduled at the begining of the simulation since this attributes depend on others, such as
DataPacketSize, which can be set by the user after calling the AquaSimFama's constructor.
The calculation of m_maxDataTxTime has been fixed.
A temporal fix for the error "Event is still running while re-scheduling" has been added.